// frontend/src/types/metric.ts
// -------------------------------------------------------------------
// Copyright (c) 2026 Ved Talmaley. All Rights Reserved.
// This project and its source code are strictly proprietary.
// Unauthorized copying, distribution, or use is strictly prohibited.
// -------------------------------------------------------------------

export interface Metric {
    metric_id: number;
    metric_name: string;
    abbreviation: string;
}